+++ R/kde.R
@@ -472,7 +472,8 @@ akde <- function(data,CTMM,VMM=NULL,debias=TRUE,weights=FALSE,smooth=TRUE,error=
   data.all <- do.call("rbind", data.all)
 
   # complete grid specification
-  grid <- kde.grid(data.all,H=H.all,axes=axes,alpha=error,res=res,dr=dr,grid=grid)
+  EXT <- extent(CTMM,level=1-error) # Gaussian extent (includes uncertainty)
+  grid <- kde.grid(data.all,H=H.all,axes=axes,alpha=error,res=res,dr=dr,grid=grid,EXT.min=EXT)
 
   # loop over individuals
   for(i in 1:n)
@@ -520,7 +521,7 @@ prepare.H <- function(H,n,axes=c('x','y'))
 # construct a grid for the density function
 # non-destructuve WRT argument 'grid'
 # non-grid arguments are merely suggestions
-kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NULL,grid=NULL)
+kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NULL,EXT.min=NULL,grid=NULL)
 {
   H <- prepare.H(H,n=length(data$t),axes=axes) # (times,dim,dim)
 
@@ -549,17 +550,6 @@ kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NUL
 
   # recycle resolution if necessary
   if("dr" %in% names(grid)) { grid$dr <- array(grid$dr,length(axes)); names(grid$dr) <- axes }
-
-  # format extents canonically
-  get.extent <- function(EXT)
-  {
-    if(class(EXT)=="Extent")
-    { EXT <- cbind( c(EXT@xmin,EXT@xmax) , c(EXT@ymin,EXT@ymax) ) }
-    else # in case of data.frame
-    { EXT <- as.matrix(EXT) }
-
-    return(EXT)
-  }
 
   # centered sequence (mu) with n steps by by
   cseq <- function(mu,n,by)
@@ -597,7 +587,7 @@ kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NUL
   else if(!is.null(grid$dr) && !is.null(grid$extent)) ### grid fully pre-specified... with possible conflict ###
   { # resolution takes presidence over extent
     dr <- grid$dr
-    EXT <- get.extent(grid$extent)
+    EXT <- as.matrix(grid$extent)
     if(isTRUE(grid$align.to.origin)) { EXT <- t((round(t(EXT)/dr+1/2)-1/2)*dr) } # assume mostly correct
 
     dEXT <- EXT[2,]-EXT[1,]
@@ -614,7 +604,7 @@ kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NUL
   else if(!is.null(grid$extent)) ### grid extent specified, but not resolution ###
   {
     # align.to.origin doesn't make sense without dr specified
-    EXT <- get.extent(grid$extent)
+    EXT <- as.matrix(grid$extent)
     dEXT <- EXT[2,]-EXT[1,]
 
     if(is.null(dr)) { dr <- dEXT/res }
@@ -655,6 +645,8 @@ kde.grid <- function(data,H,axes=c("x","y"),alpha=0.001,res=NULL,dr=NULL,EXT=NUL
     R <- get.telemetry(data,axes) # (times,dim)
 
     if(is.null(EXT)) { EXT <- rbind( apply(R-dH,2,min) , apply(R+dH,2,max) ) } # (ext,dim) }
+    colnames(EXT) <- axes
+    if(!is.null(EXT.min)) { EXT <- as.matrix(extent(list(EXT,EXT.min),level=1)) }
     dEXT <- EXT[2,]-EXT[1,]
 
     # grid center